First iteration of payload code
Still missing a bunch of things. All the stuff with reading from
the LKM still needs to be written, and we still need to steal the
radio-handling code from the ground people. The pins and some of the
timeouts/margins sorts of things still need to be decided on. But the bodies
of many of the functions are there, hopefully fleshing it out shouldn't
be terrible?